requirements-mac.txt
flask
pandas
numpy
scikit-learn
matplotlib
yfinance
alpha_vantage
tensorflow-macos
tensorflow-metal
textblob